+++ timeout/timeout.go
@@ -11,6 +11,11 @@ import (
 // ErrTimeoutExceeded is returned when an execution exceeds a configured timeout.
 var ErrTimeoutExceeded = errors.New("timeout exceeded")
 
+// Timeout is a Policy that cancels executions if they exceed a time limit. Any policies composed inside the timeout,
+// such as retries, will also be canceled. If the execution is configured with a Context, a child context will be created
+// for the execution and canceled when the Timeout is exceeded.
+//
+// This type is concurrency safe.
 type Timeout[R any] interface {
 	failsafe.Policy[R]
 }
@@ -24,7 +29,7 @@ type TimeoutBuilder[R any] interface {
 }
 
 type timeoutConfig[R any] struct {
-	timeoutDelay      time.Duration
+	timeLimit         time.Duration
 	onTimeoutExceeded func(failsafe.ExecutionDoneEvent[R])
 }
 
@@ -34,15 +39,21 @@ type timeout[R any] struct {
 	config *timeoutConfig[R]
 }
 
-// With returns a new Timeout for execution result type R and the timeoutDelay.
-func With[R any](timeoutDelay time.Duration) Timeout[R] {
-	return Builder[R](timeoutDelay).Build()
+// With returns a new Timeout for execution result type R and the timeLimit. The Timeout will cancel executions if they
+// exceed a time limit. Any policies composed inside the timeout, such as retries, will also be canceled. If the
+// execution is configured with a Context, a child context will be created for the execution and canceled when the
+// Timeout is exceeded.
+func With[R any](timeLimit time.Duration) Timeout[R] {
+	return Builder[R](timeLimit).Build()
 }
 
-// Builder returns a TimeoutBuilder for execution result type R which builds Timeouts for the timeoutDelay.
-func Builder[R any](timeoutDelay time.Duration) TimeoutBuilder[R] {
+// Builder returns a TimeoutBuilder for execution result type R which builds Timeouts for the timeLimit. The Timeout will
+// cancel executions if they exceed a time limit. Any policies composed inside the timeout, such as retries, will also be
+// canceled. If the execution is configured with a Context, a child context will be created for the execution and canceled when the Timeout
+// is exceeded.
+func Builder[R any](timeLimit time.Duration) TimeoutBuilder[R] {
 	return &timeoutConfig[R]{
-		timeoutDelay: timeoutDelay,
+		timeLimit: timeLimit,
 	}
 }